A milestone: define g.getLanguageFromAncestorAtFileNode in terms of c.getLanguage
