+++ leo/core/leoGlobals.py
@@ -36,7 +36,7 @@ import time
 import traceback
 import types
 from types import ModuleType
-from typing import Any, Generator, IO, Iterable, Optional, Sequence, Union, TYPE_CHECKING
+from typing import Any, IO, Iterable, Optional, Sequence, Union, TYPE_CHECKING
 import unittest
 import urllib
 import urllib.parse as urlparse
@@ -2617,74 +2617,12 @@ def get_directives_dict_list(p: Position) -> list[dict]:
         # No copy necessary: g.get_directives_dict does not change p.
         result.append(g.get_directives_dict(p))
     return result
-#@+node:ekr.20111010082822.15545: *3* g.getLanguageFromAncestorAtFileNode (buggy?)
+#@+node:ekr.20111010082822.15545: *3* g.getLanguageFromAncestorAtFileNode (deprecated)
 def getLanguageFromAncestorAtFileNode(p: Position) -> Optional[str]:
     """Return the language in effect at node p."""
-    v0 = p.v
-    seen: set[VNode]
-
-    # The same generator as in v.setAllAncestorAtFileNodesDirty.
-    # Original idea by Виталије Милошевић (Vitalije Milosevic).
-    # Modified by EKR.
-
-    def v_and_parents(v: VNode) -> Generator:
-        if v in seen:
-            return
-        seen.add(v)
-        yield v
-        for parent_v in v.parents:
-            if parent_v not in seen:
-                yield from v_and_parents(parent_v)
-
-    # First, see if p contains any @language directive.
-    language = g.findFirstValidAtLanguageDirective(p.b)
-    if language:
-        return language
-
-    # Passes 1 and 2: Search body text for unambiguous @language directives.
-
-    # Pass 1: Search body text in direct parents for unambiguous @language directives.
-    for p2 in p.self_and_parents(copy=False):
-        languages = g.findAllValidLanguageDirectives(p2.v.b)
-        if len(languages) == 1:  # An unambiguous language
-            return languages[0]
-
-    # Pass 2: Search body text in extended parents for unambiguous @language directives.
-    seen = set([v0.context.hiddenRootNode])
-    for v in v_and_parents(v0):
-        languages = g.findAllValidLanguageDirectives(v.b)
-        if len(languages) == 1:  # An unambiguous language
-            return languages[0]
-
-    # Passes 3 & 4: Use the file extension in @<file> nodes.
-
-    def get_language_from_headline(v: VNode) -> Optional[str]:
-        """Return the extension for @<file> nodes."""
-        if v.isAnyAtFileNode():
-            name = v.anyAtFileNodeName()
-            junk, ext = g.os_path_splitext(name)
-            ext = ext[1:]  # strip the leading period.
-            language = g.app.extension_dict.get(ext)
-            if g.isValidLanguage(language):
-                return language
-        return None
-
-    # Pass 3: Use file extension in headline of @<file> in direct parents.
-    for p2 in p.self_and_parents(copy=False):
-        language = get_language_from_headline(p2.v)
-        if language:
-            return language
-
-    # Pass 4: Use file extension in headline of @<file> nodes in extended parents.
-    seen = set([v0.context.hiddenRootNode])
-    for v in v_and_parents(v0):
-        language = get_language_from_headline(v)
-        if language:
-            return language
-
-    # Return the default language for the commander.
+    g.deprecated()
     c = p.v.context
-    return c.target_language or 'python'
+    return c.getLanguage(p)
 #@+node:ekr.20150325075144.1: *3* g.getLanguageFromPosition (deprecated)
 def getLanguageAtPosition(c: Cmdr, p: Position) -> str:
     """
