+++ leo/core/leoserver.py
@@ -2481,7 +2481,7 @@ class LeoServer:
         # Handle @killcolor and @nocolor-node when looking for language
         if c.frame.body.colorizer.useSyntaxColoring(p):
             # Get the language.
-            language = g.getLanguageFromAncestorAtFileNode(p) or c.config.getLanguage('target-language')
+            language = c.getLanguage(p)
         else:
             # No coloring at all for this node.
             language = 'plain'
